move data out of egs/, use global data zoo

+++ egs/stt_zh/run.sh
@@ -9,6 +9,7 @@ if [ $stage -le 0 ]; then
     for x in data.yaml tokenizer.yaml train.yaml test.yaml lm.yaml $text; do
         [ -f $x ] || { echo "Error: Cannot find $x"; exit 1; }
     done
+    [ -d data ] || { echo "No data dir, try 'ln -s ../../data data'"; exit 1; }
     [ -d ops ] || { echo "No ops dir, try 'ln -s ../../ops ops'"; exit 1; }
 fi
